+++ acbc_app/content/migrations/0017_remove_topictimelineentry_display_date.py
@@ -1,0 +1,15 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('content', '0016_contenttranscript'),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name='topictimelineentry',
+            name='display_date',
+        ),
+    ]