+++ frontend/src/App.jsx
@@ -41,6 +41,7 @@ import TopicEditContent from './topics/TopicEditContent';
 import TopicDetail from './topics/TopicDetail';
 import TopicContentMediaType from './topics/TopicContentMediaType';
 import TopicContentSuggestionsPage from './topics/TopicContentSuggestionsPage';
+import TopicTimelineEntryPage from './topics/timeline/TopicTimelineEntryPage';
 import ContentProfileEdit from './content/ContentProfileEdit.jsx';
 import KnowledgePathCreationForm from './knowledgePaths/KnowledgePathCreationForm';
 import KnowledgePathEdit from './knowledgePaths/KnowledgePathEdit';
@@ -159,6 +160,8 @@ const AppContent = () => {
               <Route path="create_topic" element={<ProtectedRoute><TopicCreationForm /></ProtectedRoute>} />
               <Route path="topics/:topicId/edit" element={<ProtectedRoute><TopicEdit /></ProtectedRoute>} />
               <Route path="topics" element={<TopicList />} />
+              <Route path="topics/:topicId/timeline/new" element={<ProtectedRoute><TopicTimelineEntryPage /></ProtectedRoute>} />
+              <Route path="topics/:topicId/timeline/:entryId/edit" element={<ProtectedRoute><TopicTimelineEntryPage /></ProtectedRoute>} />
               <Route path="topics/:topicId/add-content" element={<ProtectedRoute><TopicAddContent /></ProtectedRoute>} />
               <Route path="topics/:topicId/edit-content" element={<ProtectedRoute><TopicEditContent /></ProtectedRoute>} />
               <Route path="topics/:topicId" element={<ProtectedRoute><TopicDetail /></ProtectedRoute>} />